Removed Planteome from projects menu
I thought setting active=no in the project's index.md would magically keep it out of the Projects menu, but apparently the Projects menu is hardcoded in this file.

Wondering if we want to include past projects (for which we have BBOP web pages) in the menu somehow--maybe just push them to the end of the Projects menu and put "(past)" or something at the end of the name? @cmungall @kltm thoughts welcome.