remove Ben Good from people
Should we add him to emeritus/alumni? We haven't done that for people in the past other than Suzi. We don't want the People menu to get too long. Issue #21 requests addition of "alumnus" status for past projects. Could also do that for people, but will require a bit of work.