Update links and publications
this is intended to be an improvement, not perfection.

updated some out of date links (e.g. to beta.monarchinitiative.org). removed the section for internal people, which was out of date and not something we really want/need to be public. updated broken architecture image. added some newer publications. hand-edited URLs in publications to make them hyperlinks (but didn't do all of them because it's tedious).